Move attachment list logic to the component
