+++ src/frontend/app/attachment-list/attachment-list.component.html
@@ -1,9 +1,10 @@
+<input multiple type="file" class="hidden" (change)="uploadAttachments($event)" #attachmentUpload>
 <div class="flex justify-between items-center p-3">
     <span>
         <i class="fas fa-plus text-xl cursor-pointer"
            *ngIf="documentStatus !== 'END'"
            [matTooltip]="'ATTACHMENTS.add_attachments' | translate"
-           (click)="uploadAttachment()">
+           (click)="attachmentUpload.click()">
         </i>
     </span>
     <h3 class="text-center text-gray-900 m-0">
@@ -23,7 +24,7 @@
     </div>
 </div>
 
-<div *ngIf="!loadingAttachment && attachments.length === 0" class="w-full flex align-center items-center justify-center">
+<div *ngIf="!loadingAttachment && attachments.length === 0" class="w-full flex align-center items-center justify-center" (click)="attachmentUpload.click()">
     <i class="fa-solid fa-solid fa-plus text-5xl text-green-400 cursor-pointer"
        [matTooltip]="'ATTACHMENTS.add_attachments' | translate"></i>
 </div>
